% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create.varcov.maps}
\alias{create.varcov.maps}
\title{Create maps for diagonal and unconstrained varcov matrices}
\usage{
create.varcov.maps(x, elem)
}
\arguments{
\item{x}{a marssMLE object}

\item{elem}{the variance-covariance matrix: Q, R or V0}
}
\value{
a list with the factors for the diagonal and offdiagonals. NAs are elements that are not estimated (fixed).
}
\description{
The matrix is split into the diagonal and offdiag components.
factor NAs map/mask the values that are fixed and should not be estimated
}
\author{
Eli Holmes and inspired by Tim Cline's code
}
