% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfaTMB.R
\name{dfaTMB}
\alias{dfaTMB}
\title{Fit a DFA model with TMB.}
\usage{
dfaTMB(
  obs,
  NumStates = 1,
  ErrStruc = "DE",
  EstCovar = FALSE,
  Covars = NULL,
  indivCovar = FALSE,
  Dmat = NULL,
  Dfac = NULL,
  Rfac = NULL,
  logsdObs = NULL,
  logsdObsFac = NULL,
  cholCorr = NULL,
  cholFac = NULL,
  EstSE = FALSE,
  silent = TRUE
)
}
\arguments{
\item{obs}{Vector of observations n x T.}

\item{NumStates}{m}

\item{ErrStruc}{DE=diagonal and equal, UNC=unconstrained, DUE=diagonal and unequal}

\item{EstCovar}{TRUE/FALSE}

\item{silent}{Show TMB output when fitting}
}
\value{
A list with Optimization, Estimates, Fits, and AIC
}
\description{
Fit a DFA model with TMB.
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
m1.em <- MARSS(dat, model=list(R='unconstrained', m=1, tinitx=1), form='dfa', z.score=FALSE)
mym1 <- dfaTMB(obs = dat, NumStates = 1, ErrStruc='UNC')
cbind(coefficients(m1.em)$R,as.vector(mym1$Estimates$R[lower.tri(mym1$Estimates$R,diag=T)]))
cbind(coefficients(m1.em)$Z,as.vector(mym1$Estimates$Z)) 
}
\author{
Tim Cline wrote most of this while a graduate student in the Fish 507 Time Series Analysis course. Eli Holmes later modified it to replicate the MARSS(x, form="dfa") model.
}
