% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSStmb2.R
\name{estimate_marss}
\alias{estimate_marss}
\title{Internal function: MARSS parameter estimation using TMB}
\usage{
estimate_marss(MLEobj)
}
\arguments{
\item{MLEobj}{A properly formatted MARSS model as output by \code{\link[=MARSS]{MARSS()}}}
}
\value{
A list with the objective and optimization objects.
\itemize{
\item \code{obj} is the raw output from the \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} call.
\item \code{op} is the raw output from the optimization call (optim or nlminb)
}
}
\description{
Minimal error checking is done in this function.
Normal calling function is \code{\link[MARSS:MARSS]{MARSS::MARSS()}} with \code{method="TMB"}.
}
\details{
The main thing this does is
\itemize{
\item collapse the 3D fixed and free matrices into 2D
\item separate out the diag and offdiag parameter elements of R and Q
}

Restrictions
\itemize{
\item No time-varying parameters
\item V0 fixed (not estimated)
\item No B estimation
}
}
\seealso{
\code{\link[MARSS:MARSSoptim]{MARSS::MARSSoptim()}}, \code{\link[MARSS:MARSSkem]{MARSS::MARSSkem()}}
}
\author{
Eli Holmes. This function is inspired by dfaTMB.R written by Tim Cline while a graduate student in the Fish 507 Time Series Analysis course.
}
