% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSS-TMB.R
\name{MARSS.tmb}
\alias{MARSS.tmb}
\title{Main marssTMB function}
\usage{
MARSS.tmb(
  y,
  model = NULL,
  inits = NULL,
  miss.value = as.numeric(NA),
  method = "TMB",
  form = "dfa",
  fit = TRUE,
  silent = FALSE,
  control = NULL,
  ...
)
}
\arguments{
\item{y}{Vector of observations n x T.}

\item{model}{list with
\itemize{
\item R  "diagonal and equal", "unconstrained", "diagonal and unequal"
\item m the number of states (factors). default is 1
}}

\item{inits}{list of initial conditions}

\item{method}{to pass to optim call; ignored for \code{fun="nlminb"}}

\item{form}{The equation form used in the marssTMB() call. The default is "dfa".}

\item{fit}{Whether to fit the model.}

\item{silent}{Show TMB output when fitting}

\item{control}{list for the optimization function. \code{stats::nlminb()} or \code{stats::optim()}, control$fun.opt allows you to choose optim or nlminb as the optimization function. control$optim.method allows you to choose method for \code{optim()}.}
}
\value{
A list with Optimization, Estimates, Fits, and AIC
}
\description{
Fit a MARSS model with TMB + optimization. TMB does fast computation of the
likelihood gradients and this is used
with stats::nlminb()\code{or}stats::optim()` to optimize the likelihood. The
former is much faster than the later and is the default.
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
fit <- MARSS.tmb(dat, model=list(m=3))
}
\author{
Eli Holmes
}
