% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSS-TMB.R
\name{MARSS_tmb}
\alias{MARSS_tmb}
\title{Main marssTMB function}
\usage{
MARSS_tmb(
  y,
  model = NULL,
  inits = NULL,
  miss.value = as.numeric(NA),
  method = c("TMB", "BFGS.TMB", "nlminb.TMB"),
  form = c("marxss", "dfa"),
  fit = TRUE,
  silent = FALSE,
  control = NULL,
  ...
)
}
\arguments{
\item{y}{Vector of observations n x T.}

\item{model}{list with
\itemize{
\item R  "diagonal and equal", "unconstrained", "diagonal and unequal"
\item tinitx = 1 (must be)
\item (if form="dfa") m the number of states (factors). default is 1
}}

\item{inits}{list of initial conditions}

\item{miss.value}{A parameter for backcompatibility. Not used.}

\item{method}{"TMB" (default), "nlminb.TMB", or "BFGS.TMB". See details.}

\item{form}{The equation form used in the marssTMB() call. The default is "dfa".}

\item{fit}{Whether to fit the model.}

\item{silent}{Show output when fitting}

\item{control}{list for the optimization function. See \code{\link[stats:nlminb]{stats::nlminb()}} or \code{\link[stats:optim]{stats::optim()}} for the options.}

\item{...}{Extra parameters. Not used.}
}
\value{
The output list from \code{\link[=MARSStmb]{MARSStmb()}}
}
\description{
Fit a MARSS model with TMB + optimization. TMB does fast computation of the
likelihood gradients and this is used
with \code{\link[stats:nlminb]{stats::nlminb()}} or \code{\link[stats:optim]{stats::optim()}} to optimize the likelihood. The
former is much faster than the later and is the default.
}
\details{
Note in the development version of MARSS() on GitHub, you can use \code{MARSS()} as normal.
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
fit <- MARSS_tmb(dat, model=list(m=3, tinitx=1), form="dfa")
}
\author{
Eli Holmes
}
