% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marssTMB-package.R
\docType{package}
\name{marssTMB-package}
\alias{marssTMB}
\alias{marssTMB-package}
\title{marssTMB: Fits MARSS models with TMB}
\description{
Work on TMB files to fit MARSS models with TMB.
}
\author{
\strong{Maintainer}: Eli Holmes \email{eli.holmes@noaa.gov} (\href{https://orcid.org/0000-0001-9128-8393}{ORCID})

Authors:
\itemize{
  \item Eric Ward \email{eric.ward@noaa.gov} (\href{https://orcid.org/0000-0002-4359-0296}{ORCID})
  \item Timothy Cline \email{timothy.cline2@montana.edu} (\href{https://orcid.org/0000-0002-4955-654X}{ORCID})
}

}
\keyword{internal}
