% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marssTMB_dfa.R
\name{marssTMB.dfa}
\alias{marssTMB.dfa}
\title{Set up DFA Model}
\usage{
marssTMB.dfa(marssTMB.call)
}
\description{
Helper function to create and work with a DFA model sensu Zuur. The model is a MARSS model with specific assumptions (settings) on parameters. For the state equation
\eqn{\mathbf{u}}{u} is set to 0
so data and covariates should be demeaned, \eqn{\mathbf{C}}{C} and
\eqn{\mathbf{c}}{c} is set to 0, \eqn{\mathbf{Q}}{Q}  and \eqn{\mathbf{B}}{B} is set to identity.
For the observation equation, \eqn{\mathbf{a}}{a} is set to 0
so data and covariates should be demeaned,
\deqn{\mathbf{x}_{t} = \mathbf{x}_{t-1} + \mathbf{w}_t, \textrm{ where } \mathbf{w}_t \sim \textrm{MVN}(0,\mathbf{I})}{x(t) = x(t-1) + w(t), where w(t) ~ MVN(0,I)}
\deqn{\mathbf{y}_t = \mathbf{Z} \mathbf{x}_t + \mathbf{D} \mathbf{d}_t + \mathbf{v}_t, \textrm{ where } \mathbf{v}_t \sim \textrm{MVN}(0,\mathbf{R})}{y(t) = Z x(t) + D d(t) + v(t), where v(t) ~ MVN(0,R)}
\deqn{\mathbf{X}_1 \sim \textrm{MVN}(0, 5\mathbf{I}) \textrm{ or } \mathbf{X}_0 \sim \textrm{MVN}(0,  5\mathbf{I}) }{X(1) ~ MVN(0, 5I) or X(0) ~  MVN(0, 5I) }
}
\details{
The conversion functions have 2 parts
Part 1 Set up the DFA model in MARSS.marxss form
Part 2 Call MARSS.marxss to finish the set-up and checking
\subsection{arguments}{

Some arguments are common to all forms: "y" (data), "inits", "control", "method", "form", "fit", "silent", "fun.opt". See \code{\link{marssTMB()}} for
information on these arguments.
For \code{form="dfa"}, \code{m} can be passed in, e.g. \code{m=3}, to specify the number of hidden state variables (factors).

In addition to these, form="dfa" has some special arguments that can be passed in:
\itemize{
\item \code{demean}  Logical.  Default is TRUE, data will be demeaned
\item \code{z.score}  Logical.  Default is TRUE, data will be z-scored (demeaned and variance standardized to 1).
\item \code{covariates}  Covariates (\eqn{d}). No missing values allowed and must be a matrix with the same number of time steps as the data. An unconstrained \eqn{\mathbf{D}} matrix will estimated.
}
}

\subsection{model argument}{

The \code{model} argument is a list. The following details what list elements can be passed in:
\itemize{
\item \code{B}  "Identity". The standard (and default) DFA model has B="identity". However it can be "identity", "diagonal and equal", or "diagonal and unequal".
\item \code{U}  "Zero". Cannot be changed or passed in via model argument.
\item \code{Q}  "Identity". The standard (and default) DFA model has Q="identity". However, it can be "identity" or a fixed (m x m) matrix.
\item \code{Z}  Can be passed in as a (list) matrix if the user does not want a default DFA \code{Z} matrix.  There are many equivalent ways to construct a DFA \code{Z} matrix.  The default is Zuur et al.'s form with the upper triangle set to zero (see MARSS User Guide).
\item \code{A}  "zero"
\item \code{R}  Default="diagonal and equal". Can be set to "identity", "unconstrained", "diagonal and unequal", or "diagonal and equal".
\item \code{x0} "zero".
\item \code{V0} Default=diagonal matrix with 5 on the diagonal. Can be "identity" or a fixed non-zero matrix.
\item \code{tinitx} 1.  Tells MARSS that x0 is at t=1.
\item \code{m} Default=1. Can be 1 to n (the number of y time-series). Must be integer.
}
}
}
\references{
The MARSS User Guide:  Holmes, E. E., E. J. Ward, and M. D. Scheuerell (2012) Analysis of multivariate time-series using the MARSS package. NOAA Fisheries, Northwest Fisheries Science Center, 2725 Montlake Blvd E., Seattle, WA 98112
}
\author{
Eli Holmes
}
