% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{marssTMB_CheckPackageVersions}
\alias{marssTMB_CheckPackageVersions}
\title{Check TMB and Matrix versions}
\usage{
marssTMB_CheckPackageVersions(silent = TRUE)
}
\arguments{
\item{silent}{Default is TRUE. If FALSE, then gives the full instructions}
}
\description{
Helper function to provide instructions if there is a mis-match
}
