% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub3D}
\alias{sub3D}
\title{Take one 3rd element from a 3D matrix}
\usage{
sub3D(x, t = 1)
}
\arguments{
\item{x}{3D array}

\item{t}{what 3rd element to take}
}
\value{
a 2D matrix
}
\description{
Make sure that R does not turn the matrix into a vector
}
\author{
Eli Holmes
}
\keyword{internal}
