% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_marssMLE.R
\name{to.marssMLE}
\alias{to.marssMLE}
\title{Temporary helper function to make marssMLE obj}
\usage{
to.marssMLE(x)
}
\arguments{
\item{x}{list as output from \code{\link[=MARSStmb]{MARSStmb()}}: has obj1 and opt1}
}
\description{
Takes output from MARSStmb()
}
