% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_marssMLE.R
\name{to_marssMLE}
\alias{to_marssMLE}
\title{Internal function: helper function to make marssMLE obj}
\usage{
to_marssMLE(x)
}
\arguments{
\item{x}{list as output from \code{\link[=MARSStmb]{MARSStmb()}}: has obj.function and opt.output}
}
\value{
A \link[MARSS:marssMLE-class]{MARSS::marssMLE} object
}
\description{
Takes output from \code{\link[=MARSStmb]{MARSStmb()}} and makes into \link[MARSS:marssMLE-class]{MARSS::marssMLE} object.
}
