% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec}
\alias{vec}
\title{Vectorize a 2D or 3D matrix}
\usage{
vec(x)
}
\arguments{
\item{x}{2D matrix}
}
\value{
a column matrix
}
\description{
Make sure that R does not turn the matrix into a vector
}
\author{
Eli Holmes
}
\keyword{internal}
