% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSS-TMB.R
\name{MARSS_tmb}
\alias{MARSS_tmb}
\title{Main marssTMB function}
\usage{
MARSS_tmb(
  y,
  model = NULL,
  inits = NULL,
  miss.value = as.numeric(NA),
  method = "TMB",
  form = "dfa",
  fit = TRUE,
  silent = FALSE,
  control = list(fun.opt = "nlminb"),
  ...
)
}
\arguments{
\item{y}{Vector of observations n x T.}

\item{model}{list with
\itemize{
\item R  "diagonal and equal", "unconstrained", "diagonal and unequal"
\item tinitx = 1 (must be)
\item (if form="dfa") m the number of states (factors). default is 1
}}

\item{inits}{list of initial conditions}

\item{miss.value}{A parameter for backcompatibility. Not used.}

\item{method}{must be "TMB" (case sensitive)}

\item{form}{The equation form used in the marssTMB() call. The default is "dfa".}

\item{fit}{Whether to fit the model.}

\item{silent}{Show TMB output when fitting}

\item{control}{list for the optimization function. \code{\link[stats:nlminb]{stats::nlminb()}} or \code{\link[stats:optim]{stats::optim()}}, \code{control$fun.opt} allows you to choose optim or nlminb as the optimization function. \code{control$optim.method} allows you to choose method for \code{optim()}.}

\item{...}{Extra parameters. Not used.}
}
\value{
The output list from \code{\link[=MARSStmb]{MARSStmb()}}
}
\description{
Fit a MARSS model with TMB + optimization. TMB does fast computation of the
likelihood gradients and this is used
with \code{\link[stats:nlminb]{stats::nlminb()}} or \code{\link[stats:optim]{stats::optim()}} to optimize the likelihood. The
former is much faster than the later and is the default.
}
\details{
Note. Goal is to have the user call with the MARSS package as
\code{MARSS(y, method="tmb")}, which then calls \code{MARSStmb()}. This is the mimics
the behavior of \code{method="BFGS"} and \code{method="kem"} in \code{\link[MARSS:MARSS]{MARSS::MARSS()}} which
looks for a fitting function called \code{MARSSxyz}, where \code{xyz} is the method.
Further arguments for the optimization method can be passed into \code{control}.
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
fit <- MARSS_tmb(dat, model=list(m=3, tinitx=1), form="dfa")
}
\author{
Eli Holmes
}
