% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marssTMB-package.R
\docType{package}
\name{marssTMB-package}
\alias{marssTMB}
\alias{marssTMB-package}
\title{marssTMB: Fast fitting of MARSS models with TMB}
\description{
Companion to the MARSS package. Fast fitting of MARSS models with TMB. See the MARSS documentation. All the model syntax and features are the same as for the MARSS package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atsa-es.github.io/marssTMB}
  \item \url{https://github.com/atsa-es/marssTMB}
  \item Report bugs at \url{https://github.com/atsa-es/marssTMB/issues}
}

}
\author{
\strong{Maintainer}: Eli Holmes \email{eli.holmes@noaa.gov} (\href{https://orcid.org/0000-0001-9128-8393}{ORCID})

Authors:
\itemize{
  \item Eric Ward \email{eric.ward@noaa.gov} (\href{https://orcid.org/0000-0002-4359-0296}{ORCID})
  \item Timothy Cline \email{timothy.cline2@montana.edu} (\href{https://orcid.org/0000-0002-4955-654X}{ORCID})
}

}
\keyword{internal}
