% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSStmb.R
\name{MARSStmb}
\alias{MARSStmb}
\title{Internal function: Parameter estimation using TMB}
\usage{
MARSStmb(MLEobj)
}
\arguments{
\item{MLEobj}{A properly formatted MARSS model as output by \code{\link[=MARSS_tmb]{MARSS_tmb()}}}
}
\value{
A \link[MARSS:marssMLE-class]{MARSS::marssMLE} object
}
\description{
Minimal error checking is done in this function.
Normal calling function is \code{\link[=MARSS_tmb]{MARSS_tmb()}} which in
turn calls this function. Note when MARSS is updated, the
normal calling function will be \code{\link[MARSS:MARSS]{MARSS::MARSS()}}. Restrictions
}
\details{
\itemize{
\item No time-varying parameters
\item Currently only DFA models are coded up
\item x0 and V0 fixed (stochastic prior)
\item Q is fixed (not estimated)
}

This function returns \link[MARSS:marssMLE-class]{MARSS::marssMLE} object. For the \code{iter.record} element
of the object, the following are returned as a list:
\itemize{
\item \code{obj.function} is the raw output from the \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}} call.
\item \code{opt.output} is the raw output from the optimization call (optim or nlminb)
}
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

# set-up the model
mod <- MARSS_tmb(dat, model=list(m=3, tinitx=1), form="dfa", fit=FALSE)
# fit
fit <- MARSStmb(mod)

}
\seealso{
\code{\link[MARSS:MARSSoptim]{MARSS::MARSSoptim()}}, \code{\link[MARSS:MARSSkem]{MARSS::MARSSkem()}}
}
\author{
Eli Holmes. This function is inspired by dfaTMB.R written by Tim Cline while a graduate student in the Fish 507 Time Series Analysis course.
}
