% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSStmb.R
\name{MARSStmb}
\alias{MARSStmb}
\title{Parameter estimation using TMB}
\usage{
MARSStmb(MLEobj)
}
\arguments{
\item{MLEobj}{A properly formatted MARSS model as output by MARSS.tmb()}
}
\value{
A list with Optimization, Estimates, Fits, and AIC
}
\description{
Minimal error checking is done in this function.
Normal calling function is \code{MARSS.tmb()} which in
turn calls this function. Restrictions
}
\details{
\itemize{
\item No time-varying parameters
\item Currently only DFA models are coded up
\item x0 and V0 fixed (stochastic prior)
\item Q is fixed (not estimated)
}
}
\examples{
library(MARSS)
data(lakeWAplankton, package = "MARSS")
phytoplankton <- c("Cryptomonas", "Diatoms", "Greens", "Unicells", "Other.algae")
dat <- as.data.frame(lakeWAplanktonTrans) |>
  subset(Year >= 1980 & Year <= 1989) |>
  subset(select=phytoplankton) |>
  t() |>
  MARSS::zscore()

#fit with MARSS
m1.em <- MARSS(dat, model=list(R='unconstrained', m=1, tinitx=1), form='dfa', z.score=FALSE)
m1.tmb <- dfaTMB(dat, model=list(R='unconstrained', m=1))
c(m1.em$logLik, m1.tmb$logLik)
}
\author{
This function is based off of dfaTMB.R written by Tim Cline while a graduate student in the Fish 507 Time Series Analysis course. Eli Holmes later modified it to replicate the MARSS(x, form="dfa") model.
}
